feat: solve cjsToEsm.mjs
